requirements.txt
openai
redis
streamlit
fastapi
uvicorn[standard]
requests
streamlit-cookies-manager
extra-streamlit-components
Pillow